+++ src/unix/tap.rs
@@ -9,7 +9,8 @@
 // except according to those terms.
 
 use std::net::IpAddr;
-
+#[cfg(not(target_os = "windows"))]
+use std::os::fd::{AsFd, AsRawFd, BorrowedFd};
 use std::{array, io, ptr};
 
 #[cfg(not(doc))]
@@ -446,6 +447,20 @@ impl Tap {
         unsafe {
             debug_assert_eq!(libc::close(fd), 0);
         }
+    }
+}
+
+#[cfg(not(target_os = "windows"))]
+impl AsFd for Tap {
+    fn as_fd(&self) -> BorrowedFd {
+        unsafe { BorrowedFd::borrow_raw(self.fd) }
+    }
+}
+
+#[cfg(not(target_os = "windows"))]
+impl AsRawFd for Tap {
+    fn as_raw_fd(&self) -> RawFd {
+        self.fd
     }
 }